+++ akshare/__init__.py
@@ -1958,9 +1958,10 @@ amac_manager_cancelled_info # 中国证券投资基金业协会-信息公示-诚
 1.5.36 add: add stock_comment_detail_scrd_focus_em interface
 1.5.37 add: add stock_comment_detail_scrd_desire_em interface
 1.5.38 add: add stock_comment_detail_scrd_cost_em interface
+1.5.39 add: add stock_comment_detail_scrd_desire_daily_em interface
 """
 
-__version__ = "1.5.38"
+__version__ = "1.5.39"
 __author__ = "AKFamily"
 
 import sys
@@ -3141,6 +3142,7 @@ from akshare.stock_feature.stock_comment_em import (
     stock_comment_detail_zhpj_lspf_em,
     stock_comment_detail_scrd_desire_em,
     stock_comment_detail_scrd_cost_em,
+    stock_comment_detail_scrd_desire_daily_em,
 )
 
 """