+++ akshare/stock_feature/stock_comment_em.py
@@ -205,6 +205,33 @@ def stock_comment_detail_scrd_desire_em(symbol: str = "600000") -> pd.DataFrame:
     return temp_df
 
 
+def stock_comment_detail_scrd_desire_daily_em(symbol: str = "600000") -> pd.DataFrame:
+    """
+    东方财富网-数据中心-特色数据-千股千评-市场热度-日度市场参与意愿
+    https://data.eastmoney.com/stockcomment/stock/600000.html
+    :param symbol: 股票代码
+    :type symbol: str
+    :return: 市场热度-日度市场参与意愿
+    :rtype: pandas.DataFrame
+    """
+    url = f"https://data.eastmoney.com/stockcomment/api/{symbol}.json"
+    r = requests.get(url)
+    data_json = r.json()
+    date_str = data_json['ApiResults']['scrd']['desire'][0][0]['UpdateTime'].split(" ")[0].replace("/", "-")
+
+    temp_df = pd.DataFrame([data_json['ApiResults']['scrd']['desire'][2]['XData'], data_json['ApiResults']['scrd']['desire'][2]['Ydata']['PeopleNumChg'], data_json['ApiResults']['scrd']['desire'][2]['Ydata']['TotalPeopleNumChange']]).T
+    temp_df.columns = ['日期', '当日意愿下降', "五日累计意愿"]
+    temp_df['日期'] = date_str[:4] + '-' + temp_df['日期']
+    temp_df['日期'] = pd.to_datetime(temp_df['日期']).dt.date
+
+    temp_df.sort_values(['日期'], inplace=True)
+    temp_df.reset_index(inplace=True, drop=True)
+    temp_df['当日意愿下降'] = pd.to_numeric(temp_df['当日意愿下降'])
+    temp_df['五日累计意愿'] = pd.to_numeric(temp_df['五日累计意愿'])
+
+    return temp_df
+
+
 def stock_comment_detail_scrd_cost_em(symbol: str = "600000") -> pd.DataFrame:
     """
     东方财富网-数据中心-特色数据-千股千评-市场热度-市场成本
@@ -247,6 +274,9 @@ if __name__ == "__main__":
 
     stock_comment_detail_scrd_desire_em_df = stock_comment_detail_scrd_desire_em(symbol="600000")
     print(stock_comment_detail_scrd_desire_em_df)
+
+    stock_comment_detail_scrd_desire_daily_em_df = stock_comment_detail_scrd_desire_daily_em(symbol="600000")
+    print(stock_comment_detail_scrd_desire_daily_em_df)
 
     stock_comment_detail_scrd_cost_em_df = stock_comment_detail_scrd_cost_em(symbol="600000")
     print(stock_comment_detail_scrd_cost_em_df)